Added initial AppIntent to delete book
The initial app intetn has a parameter and disambiguation, and multiple
trigger phrases. There is currently no confirmation on delete.

There is an issue where Siri is not dismissed after returning, which I
am unsure of the cause currently.

When the app appears the values are updated for the App Intent / App
Shortcut, but this should also be done when the library changes.